// core-libs/core/src/model/cms.model.ts
export interface CmsComponent {
  modifiedTime?: Date;
  name?: string;
  otherProperties?: any;
  typeCode?: string;
  uid?: string;

  /**
   * Defines detailed CMS component composition
   */
  composition?: {
    /**
     * List of inner component mappings
     */
    inner?: string[];
  };

  /**
   * Style classes can be added to the CMS banner component to enhance the UX.
   * The style classes are typically derived from the (CMS) backend and should
   * match an existing CSS selector.
   *
   * The styleClasses can contain a "list" of space separated style classes.
   */
  styleClasses?: string;
}